+++ ds1820.h
@@ -10,9 +10,9 @@
 #define DS1820_CMD_READ_SCRATCHPAD   0xBE
 #define DS1820_CMD_SKIP_ROM          0xCC
 
-extern void ds1820_init(uint8_t pin);
-extern void ds1820_measure(uint8_t pin);
-extern void ds1820_write(uint8_t pin, uint8_t data);
-extern uint8_t ds1820_read_temp(uint8_t pin);
+extern void ds1820_init(uint16_t pin);
+extern void ds1820_measure(uint16_t pin);
+extern void ds1820_write(uint16_t pin, uint8_t data);
+extern uint8_t ds1820_read_temp(uint16_t pin);
 
 #endif
